Update and cleanup tests
