+++ tests/ttnn/unit_tests/operations/eltwise/test_binary_composite.py
@@ -276,46 +276,6 @@ def test_binary_div_scalar_ttnn_opt(fast_and_approximate_mode, rounding_mode, in
         (torch.Size([1, 3, 320, 384])),
     ),
 )
-def test_binary_div_no_nan_ttnn(input_shapes, device):
-    in_data1, input_tensor1 = data_gen_with_range(input_shapes, -100, 100, device)
-    in_data2, input_tensor2 = data_gen_with_range(input_shapes, -150, 150, device)
-
-    output_tensor = ttnn.div_no_nan(input_tensor1, input_tensor2)
-    golden_function = ttnn.get_golden_function(ttnn.div_no_nan)
-    golden_tensor = golden_function(in_data1, in_data2)
-
-    comp_pass = compare_pcc([output_tensor], [golden_tensor])
-    assert comp_pass
-
-
-@pytest.mark.parametrize(
-    "input_shapes",
-    (
-        (torch.Size([1, 1, 32, 32])),
-        (torch.Size([1, 1, 320, 384])),
-        (torch.Size([1, 3, 320, 384])),
-    ),
-)
-@pytest.mark.parametrize("value", [-5.1, 0.0, 10.9])
-def test_binary_div_no_nan_overload_ttnn(input_shapes, value, device):
-    in_data1, input_tensor1 = data_gen_with_range(input_shapes, -100, 100, device)
-
-    output_tensor = ttnn.div_no_nan(input_tensor1, value)
-    golden_function = ttnn.get_golden_function(ttnn.div_no_nan)
-    golden_tensor = golden_function(in_data1, value)
-
-    comp_pass = compare_pcc([output_tensor], [golden_tensor])
-    assert comp_pass
-
-
-@pytest.mark.parametrize(
-    "input_shapes",
-    (
-        (torch.Size([1, 1, 32, 32])),
-        (torch.Size([1, 1, 320, 384])),
-        (torch.Size([1, 3, 320, 384])),
-    ),
-)
 def test_binary_floor_div_ttnn(input_shapes, device):
     in_data1, input_tensor1 = data_gen_with_range(input_shapes, -350, 100, device)
     in_data2, input_tensor2 = data_gen_with_range(input_shapes, -100, 100, device)
@@ -345,139 +305,6 @@ def test_binary_floor_div_overload_ttnn(input_shapes, value, device):
 
     comp_pass = compare_pcc([output_tensor], [golden_tensor])
     assert comp_pass
-
-
-@pytest.mark.parametrize(
-    "input_shapes",
-    (
-        (torch.Size([1, 1, 32, 32])),
-        (torch.Size([1, 1, 320, 384])),
-        (torch.Size([1, 3, 320, 384])),
-    ),
-)
-def test_binary_remainder_ttnn(input_shapes, device):
-    in_data1, input_tensor1 = data_gen_with_range(input_shapes, -150, 150, device)
-    in_data2, input_tensor2 = data_gen_with_range(input_shapes, -100, 100, device)
-    output_tensor = ttnn.remainder(input_tensor1, input_tensor2)
-    golden_function = ttnn.get_golden_function(ttnn.remainder)
-    golden_tensor = golden_function(in_data1, in_data2, device=device)
-
-    comp_pass = compare_pcc([output_tensor], [golden_tensor])
-    assert comp_pass
-
-
-@pytest.mark.parametrize(
-    "shapes",
-    [
-        [[1, 1, 3, 3], [1, 1, 3, 3]],
-    ],
-)
-def test_shape_remainder(device, shapes):
-    torch.manual_seed(0)
-    high = 10
-    low = -10
-
-    torch_input_tensor_a = torch.rand(shapes[0], dtype=torch.bfloat16) * (high - low) + low
-
-    high = 9
-    low = -9
-    torch_input_tensor_b = torch.rand(shapes[1], dtype=torch.bfloat16) * (high - low) + low
-
-    golden_function = ttnn.get_golden_function(ttnn.remainder)
-    torch_output_tensor = golden_function(torch_input_tensor_a, torch_input_tensor_b, device=device)
-
-    input_tensor_a = ttnn.from_torch(
-        torch_input_tensor_a, layout=ttnn.TILE_LAYOUT, device=device, memory_config=ttnn.DRAM_MEMORY_CONFIG
-    )
-    input_tensor_b = ttnn.from_torch(
-        torch_input_tensor_b, layout=ttnn.TILE_LAYOUT, device=device, memory_config=ttnn.DRAM_MEMORY_CONFIG
-    )
-
-    output_tensor = ttnn.remainder(input_tensor_a, input_tensor_b, memory_config=ttnn.DRAM_MEMORY_CONFIG)
-    output_tensor = ttnn.to_torch(output_tensor)
-
-    assert ttnn.pearson_correlation_coefficient(torch_output_tensor, output_tensor) >= 0.999
-
-
-@pytest.mark.parametrize(
-    "input_shapes",
-    (
-        (torch.Size([1, 1, 32, 32])),
-        (torch.Size([1, 1, 320, 384])),
-        (torch.Size([1, 3, 320, 384])),
-    ),
-)
-def test_remainder_ttnn(input_shapes, device):
-    for scalar in [random.randint(-100, 100) + 0.5 for _ in range(5)]:
-        in_data1, input_tensor1 = data_gen_with_range(input_shapes, -150, 150, device)
-        output_tensor = ttnn.remainder(input_tensor1, scalar)
-        golden_function = ttnn.get_golden_function(ttnn.remainder)
-        golden_tensor = golden_function(in_data1, scalar, device=device)
-
-        comp_pass = compare_pcc([output_tensor], [golden_tensor])
-        assert comp_pass, f"Failed for scalar={scalar}"
-
-
-@pytest.mark.parametrize(
-    "input_shapes",
-    (
-        (torch.Size([1, 1, 32, 32])),
-        (torch.Size([1, 1, 320, 384])),
-        (torch.Size([1, 3, 320, 384])),
-    ),
-)
-def test_binary_fmod_ttnn(input_shapes, device):
-    in_data1, input_tensor1 = data_gen_with_range(input_shapes, -150, 150, device)
-    in_data2, input_tensor2 = data_gen_with_range(input_shapes, -100, 100, device)
-
-    output_tensor = ttnn.fmod(input_tensor1, input_tensor2)
-    golden_function = ttnn.get_golden_function(ttnn.fmod)
-    golden_tensor = golden_function(in_data1, in_data2, device=device)
-
-    comp_pass = compare_pcc([output_tensor], [golden_tensor])
-    assert comp_pass
-
-
-@pytest.mark.parametrize(
-    "input_shapes",
-    (
-        (torch.Size([1, 1, 32, 32])),
-        (torch.Size([1, 1, 320, 384])),
-        (torch.Size([1, 3, 320, 384])),
-    ),
-)
-# Input with more than two decimal places experience precision loss in bfloat16. use FP32 for better precision.
-def test_binary_fmod_decimal_ttnn(input_shapes, device):
-    in_data1 = torch.randn(input_shapes, dtype=torch.float32) * 9
-    input_tensor1 = ttnn.Tensor(in_data1, ttnn.float32).to(ttnn.TILE_LAYOUT).to(device)
-    in_data2 = torch.rand(input_shapes, dtype=torch.float32) - 2
-    input_tensor2 = ttnn.Tensor(in_data2, ttnn.float32).to(ttnn.TILE_LAYOUT).to(device)
-    output_tensor = ttnn.fmod(input_tensor1, input_tensor2)
-    golden_function = ttnn.get_golden_function(ttnn.fmod)
-    golden_tensor = golden_function(in_data1, in_data2, device=device)
-
-    output_torch = ttnn.to_torch(output_tensor)
-    assert torch.allclose(output_torch, golden_tensor, rtol=5e-2, atol=1e-5)
-
-
-@pytest.mark.parametrize(
-    "input_shapes",
-    (
-        (torch.Size([1, 1, 32, 32])),
-        (torch.Size([1, 1, 320, 384])),
-        (torch.Size([1, 3, 320, 384])),
-    ),
-)
-def test_fmod_ttnn(input_shapes, device):
-    for scalar in [random.randint(-100, 100) + 0.5 for _ in range(5)]:
-        in_data1, input_tensor1 = data_gen_with_range(input_shapes, -150, 150, device)
-
-        output_tensor = ttnn.fmod(input_tensor1, scalar)
-        golden_function = ttnn.get_golden_function(ttnn.fmod)
-        golden_tensor = golden_function(in_data1, scalar, device=device)
-
-        comp_pass = assert_with_ulp(golden_tensor, output_tensor, 1)
-        assert comp_pass, f"Failed for scalar={scalar}"
 
 
 @pytest.mark.parametrize(
